requirements.txt
bottle==0.12.13
certifi==2017.11.5
chardet==3.0.4
idna==2.6
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
